# Emit attributes slices as their json representation