@dmathieu 2024-04-05
> Do we consider this breaking? An alternative would be to have a new EmitJSON (or other) function.

This may be considered a breaking change, but I don't know if a new `EmitJSON` method would change anything. `Emit()` is only called by exporters, and they would all have to start calling that new method (and have conditions to check which one to call).

It seems to me the breaking change here is in the format attributes are exporter in. I'm not sure we have a trivial solution though.