@dashpole 2024-04-05
It looks like it is only used by zipkin and prometheus.  Zipkin wouldn't be affected by the change regardless, since it already uses json for complex types: https://github.com/open-telemetry/opentelemetry-go/blob/82b49b448731e3ca3d88d7421d03768e417f97f4/exporters/zipkin/model.go#L172

Prometheus isn't stable, so it can make the change.