@dashpole 2024-04-12
I'm still concerned that this will be breaking for users.  How do people feel about adding add a new EmitJSON function with this behavior, and leaving the existing Emit function as-is?