@dmathieu 2024-04-12
How does the exporter decide which method it should call?