@dashpole 2024-04-12
We should recommend that exporters call EmitJSON.  We could even consider deprecating (but not removing) Emit().