@dmathieu 2024-04-22
While I understand the breaking change concern, I don't think `EmitJSON` is a good solution here.
In order to avoid having to decide whether we should call `Emit()`, or `EmitJSON()` depending on the attribute type, we'd have to call `EmitJSON` every time.

Except if we emit a string as json, well get `"string"` while today, we emit `string`. So we would be telling folks they're getting a JSON representation, when that's not accurate.