@dashpole 2024-04-22
Thanks.  I do think this change is worth making, especially since it brings us into compliance with a part of a stable specification.